backend/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
asn1crypto==1.3.0
attrs==19.3.0
autobahn==20.2.2
Automat==20.2.0
Babel==2.8.0
bcrypt==3.1.7
billiard==3.6.2.0
cachetools==4.0.0
celery==4.3.0
certifi==2018.11.29
cffi==1.14.0
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
cs==2.7
daphne==2.4.1
dateparser==0.7.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-appconf==1.0.2
django-celery==3.2.2
django-cors-headers==2.4.0
django-crequest==2018.5.11
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-imagekit==4.0.2
django-import-export==1.1.0
django-ipware==2.1.0
django-redis-cache==2.1.0
django-rest-swagger==2.2.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
dwebsocket==0.5.11
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
et-xmlfile==1.0.1
flower==0.9.3
future==0.17.1
gevent==1.4.0
google-auth==1.11.2
greenlet==0.4.15
gunicorn==19.9.0
httplib2==0.12.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jira==2.0.0
jmespath==0.9.5
kombu==4.5.0
ldap3==2.5.2
Markdown==3.0.1
MarkupSafe==1.1.0
multi-key-dict==2.0.3
mysqlclient==1.3.14
oauthlib==3.1.0
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.5.12
paramiko==2.7.1
pbr==5.4.4
pilkit==2.0
Pillow==5.3.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.1
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyotp==2.2.7
python-dateutil==2.8.1
python-dotenv==0.10.3
python-ldap==3.1.0
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
qrcode==6.1
redis==3.2.0
regex==2020.2.20
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.12.1
tornado==5.1.1
Twisted==19.10.0
txaio==20.1.1
tzlocal==2.0.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.3
uWSGI==2.0.17.1
vine==1.3.0
websocket-client==0.57.0
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmltodict==0.12.0
xpinyin==0.5.6
zmail==0.2.3
zope.interface==4.7.1
kombu==4.5.0
celery==4.3.0
redis==3.2.0
numpy==1.18.1